% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{datos_pedido}
\alias{datos_pedido}
\title{Extraer datos de clientes}
\usage{
datos_pedido(token, hostname, id_cliente, tam_grupo = 100, verbose = FALSE)
}
\arguments{
\item{token}{Clave de tu webservice}

\item{hostname}{Nombre de tu página web. Por ejemplo, si la URL es https://www.prueba.com/, el hostname es www.prueba.com.}

\item{id_cliente}{Vector con el o los ids de clientes a extraer.}

\item{tam_grupo}{El número de clientes de los que se pedirá información con cada petición para evitar saturar el servidor.}

\item{verbose}{Indica si la función debería indicar cada paso de la extracción de datos.}
}
\value{
Un dataframe con todos los datos extraidos.
}
\description{
Esta función te permite extraer los datos de cualquier endpoint de  prestashop mediante su API.
Para ello, simplemente debes indicar el token, el hostname y los filtros que quieras aplicar.
}
