% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presta_data.R
\name{presta_data}
\alias{presta_data}
\title{Extract order data from Prestashop}
\usage{
presta_data(
  token,
  hostname,
  endpoint = "orders",
  id_column = "id",
  ids = NULL,
  group_dim = 100,
  verbose = FALSE
)
}
\arguments{
\item{token}{Webservice Key.}

\item{hostname}{Hotname of your website. For example, si your URL is https://www.test.com/, the hostname is www.test.com. On the other hand, if your URL is https://test.com/ your hostname is test.com.}

\item{endpoint}{Indicates the table you want to retrieve data from. The webservice should have permissions.}

\item{ids}{A vector including the ids that you want to extract. To extract all the data, set parameter to NULL.}

\item{group_dim}{The number of rows that will be retrieved in each API call. Set as default if you will retrieve a lot of data.}

\item{verbose}{Whether you want to see the process of extraction.}
}
\value{
A dataframe with all extracted data
}
\description{
This function enables you to extract data from the Orders table of Prestashop.
To do so, you just need to indicate the hostname, token and the orders of which you want to extract data from.
}
