% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presta_extract_ids.R
\name{presta_extract_ids}
\alias{presta_extract_ids}
\title{Extract ids from Prestashop Table}
\usage{
presta_extract_ids(
  token,
  hostname,
  endpoint = "orders",
  verbose = FALSE,
  date_min = NULL,
  date_max = NULL,
  date_column = "date_add"
)
}
\arguments{
\item{token}{Webservice Key.}

\item{hostname}{Hotname of your website. For example, si your URL is https://www.test.com/, the hostname is www.test.com. On the other hand, if your URL is https://test.com/ your hostname is test.com.}

\item{endpoint}{Indicates the table you want to retrieve data from. The webservice should have permissions.}

\item{verbose}{Whether you want to see the process of extraction.}

\item{date_min}{Minimum date to get data from. Not required. If used, date_max is also required.}

\item{date_max}{Maximum date to get data from. Not required. If used, date_min is also required.}

\item{date_add}{Column to use for filtering data by date. Defaul is date_add.}
}
\value{
A vector with all the ids
}
\description{
This function enables you to extract the ids from a selected Prestashop table.
To do so, you just need to indicate the hostname, token and the table from where you want to extract ids.
}
